+++ modules/camera/include/ifm3d/camera/camera.h
@@ -24,31 +24,23 @@
 #include <string>
 #include <vector>
 #include <ifm3d/contrib/json.hpp>
+#include <ifm3d/camera/camera_export.h>
 
 using json = nlohmann::json;
 
-// Annoying stuff for windows -- makes sure clients can import these functions
-#ifndef IFM3D_DLL_DECL
-# if defined(_WIN32) && !defined(__CYGWIN__)
-#   define IFM3D_DLL_DECL  __declspec(dllimport)
-# else
-#   define IFM3D_DLL_DECL
-# endif
-#endif
-
 namespace ifm3d
 {
-  extern IFM3D_DLL_DECL const std::string DEFAULT_IP;
-  extern IFM3D_DLL_DECL const std::uint16_t DEFAULT_XMLRPC_PORT;
-  extern IFM3D_DLL_DECL const int DEFAULT_PCIC_PORT;
-  extern IFM3D_DLL_DECL const std::string DEFAULT_PASSWORD;
-  extern IFM3D_DLL_DECL const int MAX_HEARTBEAT;
+  extern IFM3D_CAMERA_EXPORT const std::string DEFAULT_IP;
+  extern IFM3D_CAMERA_EXPORT const std::uint16_t DEFAULT_XMLRPC_PORT;
+  extern IFM3D_CAMERA_EXPORT const int DEFAULT_PCIC_PORT;
+  extern IFM3D_CAMERA_EXPORT const std::string DEFAULT_PASSWORD;
+  extern IFM3D_CAMERA_EXPORT const int MAX_HEARTBEAT;
 
-  extern IFM3D_DLL_DECL const int DEV_O3D_MIN;
-  extern IFM3D_DLL_DECL const int DEV_O3D_MAX;
-  extern IFM3D_DLL_DECL const int DEV_O3X_MIN;
-  extern IFM3D_DLL_DECL const int DEV_O3X_MAX;
-  extern IFM3D_DLL_DECL const std::string ASSUME_DEVICE;
+  extern IFM3D_CAMERA_EXPORT const int DEV_O3D_MIN;
+  extern IFM3D_CAMERA_EXPORT const int DEV_O3D_MAX;
+  extern IFM3D_CAMERA_EXPORT const int DEV_O3X_MIN;
+  extern IFM3D_CAMERA_EXPORT const int DEV_O3X_MAX;
+  extern IFM3D_CAMERA_EXPORT const std::string ASSUME_DEVICE;
 
   /**
    * Software interface to an ifm 3D camera
@@ -290,6 +282,16 @@ namespace ifm3d
      * (i.e., no edit session created on the camera)
      */
     virtual std::string DeviceParameter(const std::string& key);
+
+    /**
+     * Delivers the trace log from the camera
+     * A session is not required to call this function.
+     *
+     * @return A `vector' of `std::string' for each entry in the tracelog
+     *
+     * @throw ifm3d::error_t upon error
+     */
+     virtual std::vector<std::string> TraceLogs(int count);
 
     /**
      * Delivers basic information about all applications stored on the device.